Fix cowboy dependancy issue
As discussed here, https://github.com/phoenixframework/phoenix/issues/3119, in some cases, with Phoenix 1.3, the endpoint will fail to start with `** (EXIT) "plug_cowboy dependency missing"` if `cowboy` instead of `plug_cowboy` is listed as a dep. I ran into that issue with this example. The proposed change, recommended in the link above, fixed it for me.